# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "nexus", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/auth/google-callback": {
      "post": {
        "tags": ["google_oauth"],
        "summary": "Google Callback Api",
        "description": "Handle Google OAuth callback from frontend\nThis is maintained for backward compatibility but not used in the new flow",
        "operationId": "google_oauth-google_callback_api",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GoogleCallbackRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/login/google": {
      "get": {
        "tags": ["google_oauth"],
        "summary": "Google Login",
        "description": "Initiate Google OAuth2 authentication flow\nThis endpoint redirects to Google's login page",
        "operationId": "google_oauth-google_login",
        "parameters": [
          {
            "name": "extension_callback",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Extension Callback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/login/google/callback": {
      "get": {
        "tags": ["google_oauth"],
        "summary": "Google Callback",
        "description": "Handle the callback from Google OAuth\nThis endpoint is called by Google after the user has logged in",
        "operationId": "google_oauth-google_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "State"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat/completions": {
      "post": {
        "tags": ["chat"],
        "summary": "Create Chat Completion",
        "description": "Create chat completion compatible with Vercel AI SDK.\n\nSupports Data Stream Protocol format for frontend integration.",
        "operationId": "chat-create_chat_completion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_chat-create_chat_completion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/chat/models": {
      "get": {
        "tags": ["chat"],
        "summary": "List Available Models",
        "description": "\u5217\u51fa\u53ef\u7528\u7684 AI \u6a21\u578b",
        "operationId": "chat-list_available_models",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_chat-create_chat_completion": {
        "properties": {
          "messages": {
            "items": { "type": "object" },
            "type": "array",
            "title": "Messages",
            "description": "Chat messages in OpenAI format"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model to use",
            "default": "or-gemini-2.5-flash-preview-05-20"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "Whether to stream the response",
            "default": true
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "description": "Sampling temperature",
            "default": 0.7
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "description": "Maximum tokens to generate",
            "default": 2000
          }
        },
        "type": "object",
        "required": ["messages"],
        "title": "Body_chat-create_chat_completion"
      },
...
      "GoogleCallbackRequest": {
        "properties": {
          "token": { "type": "string", "title": "Token" },
          "user_info": { "type": "object", "title": "User Info" }
        },
        "type": "object",
        "required": ["token", "user_info"],
        "title": "GoogleCallbackRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}